requirements.txt
magicwand
requests
pyinstaller